arch: migrate telegram_bot.integrations.cache + bge_m3_query_bundle + vectorizers to src.runtime / src.services (#2049 slice 4)
Fourth slice of the reverse-layering migration tracked under #1948 / #2047 / #2049. Builds on #2096 (graph.state), #2099 (graph.config + embeddings), #2100 (qdrant + metrics). Moves the 5-tier Redis CacheLayerManager and its two helper modules out of telegram_bot/ so src/api/main.py no longer imports CacheLayerManager from the bot package.

What this PR does:

- Adds src/runtime/integrations/cache.py (canonical home for CacheLayerManager, CACHE_VERSION, SEMANTIC_CACHE_VERSION, DEFAULT_TTLS, BGE_M3_QUERY_BUNDLE_MODEL_NAME, plus the _create_semantic_cache / _create_embed_cache / _hash / _normalize_query_for_cache / _redact_redis_credentials private helpers). The new module imports observability (get_client, observe) from src.observability and the BGE-M3 query bundle / vectorizer types from src.services.* instead of telegram_bot.*, satisfying the layering contract.

- Adds src/services/bge_m3_query_bundle.py (canonical home for BgeM3QueryVectorBundle and the BGE_M3_QUERY_BUNDLE_* constants + make_bge_m3_query_bundle_key_material helper). Standalone module, only stdlib deps, copied verbatim.

- Adds src/services/vectorizers.py (canonical home for UserBaseVectorizer and BgeM3CacheVectorizer). Lazy import inside BgeM3CacheVectorizer.embed_many now resolves BGEM3Client from src.services.bge_m3_client (the canonical client home introduced in earlier slices).

- Reduces telegram_bot/integrations/cache.py, telegram_bot/services/bge_m3_query_bundle.py, and telegram_bot/services/vectorizers.py to thin re-export shims so every existing caller (40+ test files, src/api, scripts/, telegram_bot internals) keeps working without churn. The cache shim also re-exports the private helpers (_create_semantic_cache, _create_embed_cache, _hash, _normalize_query_for_cache, _redact_redis_credentials) to keep the import surface stable for the unit suite.

- Switches src/api/main.py:40 to import CacheLayerManager from src.runtime.integrations.cache.

- Shrinks tests/data/known_layering_violations.json from 2 entries to 1 for src/api/main.py: telegram_bot.integrations.cache is now compliant. The remaining entry (telegram_bot.graph.graph) is the slice-5 target.

Test contract path updates (necessary because canonical home moved):

- tests/unit/integrations/test_cache_layers.py: 17 patch targets and the caplog logger name updated from telegram_bot.integrations.cache.* to src.runtime.integrations.cache.* (the patches reach into module-internal helpers that the canonical implementation now owns; patching the shim's namespace would not affect the canonical module's resolution).

- tests/contract/test_error_contract.py: ERROR_SPAN_ALLOWLIST gains src/runtime/integrations/cache.py alongside the legacy shim path.

What this PR does NOT do (out of scope, follow-up work):

- The remaining src/api/main.py reverse import (telegram_bot.graph.graph) flows back into telegram_bot. It is the slice-5 target and depends on the cache slice landing first. Issue #2049 stays open until the allowlist is empty and a contract test asserts allowlist == [].

- Mini App separability verification is not part of this slice.

Validation:

- uv run --frozen pytest tests/contract/test_layering_no_telegram_bot_imports_contract.py tests/contract/test_error_contract.py -q --no-cov: 292 passed (allowlist drift detector picks up the shrunk JSON; error-contract picks up the new canonical path).

- uv run --frozen pytest tests/unit/services/ tests/unit/graph/ tests/unit/integrations/ tests/contract/ -q --no-cov: 3408 passed, 5 skipped, 3 xfailed (back-compat shims work for every existing caller).

- ruff check + ruff format clean on touched files.

Refs #2049 #2047 #1948
