+++ src/api/main.py
@@ -35,9 +35,9 @@ _LANGFUSE_TRACE_ID_RE = re.compile(r"^[0-9a-f]{32}$")
 async def lifespan(app: FastAPI):
     """Initialize and teardown pipeline services."""
     from src.runtime.graph.config import GraphConfig
+    from src.runtime.integrations.cache import CacheLayerManager
     from src.runtime.services.qdrant import QdrantService
     from telegram_bot.graph.graph import build_graph
-    from telegram_bot.integrations.cache import CacheLayerManager
 
     cfg = GraphConfig.from_env()
